Minor adjustments after rebase
